refactor: unify top bar for portrait/object modes (interim)
Portrait and object modes now share a single TopCameraBar; landscape mode
is intentionally left untouched.

- Add optional `badge` slot to TopCameraBar (lib/screen/camera/widgets/
  top_camera_bar.dart).
- Extract portrait badge into PortraitBadge widget (portrait_badge.dart)
  and pass it through `badge` in portrait mode.
- Remove `_buildPortraitTopBar()` and the now-unused PortraitTopBar
  widget file. Portrait mode now exposes flash/timer toggles like
  object mode; ROI lock remains object-mode only (no meaning for
  portrait subject tracking).
- Coaching message UI was already unified (CoachingSpeechBubble) across
  modes -- no change there.

No behavior change for landscape mode. flutter analyze passes with no
new errors.
